prefer Double to Int for most values in axle.visualize (SVG supports it, awt now .toInt everything)

+++ axle-visualize/src/main/scala/axle/visualize/element/DataPoints.scala
@@ -7,4 +7,4 @@ case class DataPoints[X, Y](
   scaledArea: ScaledArea2D[X, Y],
   data: Set[(X, Y)],
   colorStream: Stream[Color],
-  pointDiameter: Int)
+  pointDiameter: Double)
